+++ src/routes/index.tsx
@@ -1,24 +1,11 @@
-import { TOKEN_KEY } from '@/service/constants'
 import heladeriaService from '@/service/heladeria-service'
+import { onBeforeLoad, onErrorRoute } from '@/utils/routes'
 import Home from '@/views/Home'
-import { createFileRoute, redirect } from '@tanstack/react-router'
+import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/')({
   component: Home,
-  beforeLoad: () => {
-    const isLoggedIn = localStorage.getItem(TOKEN_KEY) !== null
-    if (!isLoggedIn) {
-      throw redirect({
-        to: '/login',
-        search: {
-          // Use the current location to power a redirect after login
-          // (Do not use `router.state.resolvedLocation` as it can
-          // potentially lag behind the actual current location)
-          redirect: location.href,
-        },
-      })
-    }
-  },
+  beforeLoad: onBeforeLoad,
   validateSearch: (search) => {
     return search as {
       busqueda?: string
@@ -28,4 +15,5 @@ export const Route = createFileRoute('/')({
     return { busqueda: params.search.busqueda }
   },
   loader: async (params) => heladeriaService.buscarHeladerias(params.deps.busqueda),
+  onError: onErrorRoute,
 })
