+++ src/views/Login/index.tsx
@@ -2,7 +2,6 @@ import Button from '@/components/Button'
 import Card from '@/components/Card'
 import Icon from '@/components/Icon'
 import Input from '@/components/Input'
-import Label from '@/components/Label'
 import { TOKEN_KEY } from '@/service/constants'
 import { loginUser } from '@/service/usuario-service'
 import { useRouter } from '@tanstack/react-router'
